rename CmsPageCategoriesBreadcrumbTree to BreadcrumbTree and move to QueryResult namespace
